feat: preload feed on home page
